Use loadtest4j 0.16.0

+++ src/test/java/org/loadtest4j/examples/java/driver/DummyResponseTime.java
@@ -6,7 +6,7 @@ import java.time.Duration;
 
 class DummyResponseTime implements DriverResponseTime {
     @Override
-    public Duration getPercentile(int i) {
+    public Duration getPercentile(double i) {
         return Duration.ZERO;
     }
 }
